+++ AIO.Contracts/Interfaces/Services/Permissions/IPermissionService.cs
@@ -8,7 +8,7 @@ namespace AIO.Contracts.IServices.Services.Permissions
         public Task<IHolderOfDTO> GetAllAsync();
         public Task<IHolderOfDTO> GetByIdAsync(long id);
         public Task<IHolderOfDTO> GetPermissionByRoleId(string RoleId);
-        public Task<Dictionary<string, List<long>>> GetPermissionByRolesName(List<string> RolesName);
+        public Task<Dictionary<string, List<int>>> GetPermissionByRolesName(List<string> RolesName);
         public Task<IHolderOfDTO> SaveAsync(List<RolePermissionSetterDTO> RolePermissionSetterDTO);
         public IHolderOfDTO Update(RolePermissionSetterDTO RolePermissionSetterDTO);
         public IHolderOfDTO Delete(long id);
